requirements.txt
python-telegram-bot==20.0
PyPDF2
pytesseract
pdf2image
httpx
